Unpin apt-get versions in Dockerfile
Obviously pinning is the ideal, as each build of the image will use the
same versions of packages, however, when Ubuntu patches a package it
removes the old version. This means the build will fail as the old
version of the package is no longer available.

As the patches are only minor bug and security fixes, in our case I
think it's reasonable to unpin.