Update tests for compatibility with RunCommand
A recent MOOSE update meant that the way we were using the 'RunCommand'
test type stopped working. It seems like the first argument in the
command string is now taken as the executable, and all subsequent parts
of the string are passed as arguments to that executable. We were
therefore getting errors when chaining commands together with '&&'.

We can work around this by prefixing our one-line scripts with 'bash
-c'.