Prefer Docker image tag to 'latest' in CI workflow
Using the 'latest' tag means that we must update the "production"
version of the docker image _before_ we run a CI job using that image.
This can create a bit of a deadlock where, if we update the image in
some PR, we must update the 'latest' tag on the Docker image then run
the tests on the PR. If the PR fails, we must fix that PR before we can
do anything else.

Using the specific tags resolves this problem. Once we have merged the
PR, we can then (manually, for now) update the image tag to 'latest'.