Always initialise an OpenFOAM solver
There seems to be a bit of weirdness when not initialising a solver,
where OpenFOAM doesn't cope well with destruction of Time objects. I'd
get errors saying 'Cannot access element zero of list' (or similar) when
the OpenFOAM object trees were being destructed.

Workaround this by always intialising a solver.